requirements.txt
flask==3.0.2
yt-dlp==2024.3.10
gunicorn==23.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
youtube-search-python==1.6.6
requests==2.31.0